requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiogram-dialog==2.0.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
black==24.2.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
frozenlist==1.4.1
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
gspread==6.0.2
httplib2==0.22.0
idna==3.6
Jinja2==3.1.3
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.0.5
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
StrEnum==0.4.15
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
